fix: spdlog compile-time format string issues
- Use spdlog::fmt_lib::runtime() to wrap format strings
- This allows runtime format strings with spdlog 1.17+ / fmt 11.x
- Remove SPDLOG_COMPILED_LIB definition (header-only mode)
- Cast VkFormat to int for logging (no fmt formatter)

+++ src/renderer/Swapchain.cpp
@@ -217,7 +217,7 @@ bool Swapchain::createSwapchain(int width, int height) {
     vkGetSwapchainImagesKHR(m_device, m_swapchain, &imageCount, m_images.data());
 
     LOG_INFO("Swapchain created: {}x{}, {} images, format {}", 
-              m_extent.width, m_extent.height, imageCount, m_format);
+              m_extent.width, m_extent.height, imageCount, static_cast<int>(m_format));
     
     return true;
 }
